media-source/MediaSource-readyState.html

<!doctype html>
<html>
<head>
  <meta charset='utf-8'>
  <title>insertAdjacentHTML</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div>
    <video id='v' controls autoplay width='320' height='240'></video>
</div>
<div id="log"></div>

<script>
test(function() {
    var mediaSource = new MediaSource();
    assert_equals(mediaSource.readyState, 'closed');
}, 'MediaSource#readyState : Initial state');
</script>
</body>
</html>
